+++ cmd/my-cli/root.go
@@ -51,6 +51,7 @@ Run "my-cli <command> --help" for details on a specific command.`,
 	rootCmd.AddCommand(newOpenaiChatTestCmd())
 	rootCmd.AddCommand(newFalconCisExportCmd())
 	rootCmd.AddCommand(newNexusExportCmd())
+	rootCmd.AddCommand(newIPLookupCmd())
 
 	return rootCmd
 }